// packages/react-client/context/evm/types/ConnectionStatus.ts
import { WalletClientWithAccount } from '@nadohq/client';
import { Address } from 'viem';
import { Connector } from 'wagmi';

/**
 * All possible connection states
 */
export type ConnectionStatus =
  | {
      type: 'initializing' | 'disconnected' | 'reconnecting' | 'connecting';
      address: undefined;
      walletClient: undefined;
      connector: Connector | undefined;
    }
  | {
      type: 'connected';
      address: Address;
      walletClient: WalletClientWithAccount | undefined;
      connector: Connector;
    };